Revert "Set dtype to float32."
This reverts commit 4f18f6bba22f234647f96e1492afe0eadf6d93d4.

+++ incomplete_cooperative/protocols.py
@@ -17,7 +17,7 @@ if TYPE_CHECKING:  # pragma: no cover
     from .run.model import ModelInstance
 
 Player = int
-Value = np.float32
+Value = np.float64
 ValueIn = Value | float | int
 Values = np.ndarray[Any, np.dtype[Value]]
